+++ tests/forms-schema/form-schema-conditional-logic.test.ts
@@ -1010,4 +1010,139 @@ describe('Conditional Predicates', () => {
       '"elements[2].conditionallyShowPredicates[1].value" must be a number',
     )
   })
+  test('should allow REPEATABLESET conditional predicate type', () => {
+    console.log(formSchema)
+    const result = formSchema.validate(
+      {
+        id: 1,
+        name: 'conditionally show element via number input',
+        description: 'string',
+        formsAppEnvironmentId: 1,
+        formsAppIds: [1],
+        organisationId: 'ORGANISATION_00000000001',
+        postSubmissionAction: 'FORMS_LIBRARY',
+        isMultiPage: false,
+        elements: [
+          {
+            id: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b7',
+            name: 'RepeatableSet',
+            label: 'Repeatable Set',
+            type: 'repeatableSet',
+            required: false,
+            elements: [
+              {
+                id: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b1',
+                name: 'Numbers',
+                label: 'Numbers',
+                type: 'number',
+                required: false,
+                minNumber: 1,
+                maxNumber: 6,
+                defaultValue: 3,
+              },
+            ],
+          },
+          {
+            id: '8e4d819b-97fa-438d-b613-a092d38c3b23',
+            name: 'Text',
+            label: 'Text',
+            type: 'text',
+            required: false,
+            defaultValue: 'text',
+            isDataLookup: false,
+            isElementLookup: false,
+            readOnly: false,
+            conditionallyShow: true,
+            requiresAllConditionallyShowPredicates: false,
+            conditionallyShowPredicates: [
+              {
+                elementId: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b7',
+                type: 'REPEATABLESET',
+                repeatableSetPredicate: {
+                  elementId: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b1',
+                  type: 'BETWEEN',
+                  min: 2,
+                  max: 5,
+                },
+              },
+            ],
+          },
+        ],
+        isAuthenticated: true,
+      },
+
+      {
+        abortEarly: false,
+      },
+    )
+    expect(result.error).toBe(undefined)
+  })
+
+  test('should NOT allow REPEATABLESET conditional predicate type in repeatableSetPredicate', () => {
+    console.log(formSchema)
+    const result = formSchema.validate(
+      {
+        id: 1,
+        name: 'conditionally show element via number input',
+        description: 'string',
+        formsAppEnvironmentId: 1,
+        formsAppIds: [1],
+        organisationId: 'ORGANISATION_00000000001',
+        postSubmissionAction: 'FORMS_LIBRARY',
+        isMultiPage: false,
+        elements: [
+          {
+            id: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b7',
+            name: 'RepeatableSet',
+            label: 'Repeatable Set',
+            type: 'repeatableSet',
+            required: false,
+            elements: [
+              {
+                id: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b1',
+                name: 'Numbers',
+                label: 'Numbers',
+                type: 'number',
+                required: false,
+                minNumber: 1,
+                maxNumber: 6,
+                defaultValue: 3,
+              },
+            ],
+          },
+          {
+            id: '8e4d819b-97fa-438d-b613-a092d38c3b23',
+            name: 'Text',
+            label: 'Text',
+            type: 'text',
+            required: false,
+            defaultValue: 'text',
+            isDataLookup: false,
+            isElementLookup: false,
+            readOnly: false,
+            conditionallyShow: true,
+            requiresAllConditionallyShowPredicates: false,
+            conditionallyShowPredicates: [
+              {
+                elementId: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b7',
+                type: 'REPEATABLESET',
+                repeatableSetPredicate: {
+                  elementId: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b1',
+                  type: 'REPEATABLESET',
+                },
+              },
+            ],
+          },
+        ],
+        isAuthenticated: true,
+      },
+
+      {
+        abortEarly: false,
+      },
+    )
+    expect(result.error?.message).toBe(
+      '"elements[1].conditionallyShowPredicates[0].repeatableSetPredicate.type" must be one of [OPTIONS, NUMERIC, VALUE, BETWEEN]',
+    )
+  })
 })