+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - CI job for building docs
+- `REPEATABLESET` type to conditional predicates
 
 ## [4.0.0] - 2023-08-17